refactor(util): 책임에 따라 Util 클래스를 이동 및 수정
- NumberGenerator를 test source로 이동
- MovementUtil가 하던 일을 Car안에서 하도록 이동
- RandomUtil를 만들어 랜덤값에 대한 일을 하도록 함
