+++ docs/README.md
@@ -187,10 +187,19 @@
 - [x] View에서만 사용되는 에러 메시지를 왜 Domain 패키지에 위치한걸까요?
     - 아직 view와 domain의 경계를 구분하지 못하고 있어서 포함시킨 것 같습니다.
     - View에서 발생하는 예외 처리에 사용할 ViewErrorMessage와 domain에서 발생하는 예외 처리에 사용할 DomainErrorMessage로 분리.
-- [ ] Util 클래스는 어떤 책임을 가져야 할까요?
-- [ ] 테스트에서만 사용되는 클래스는 test source set에 선언해주면 어떨까요?
+- [x] Util 클래스는 어떤 책임을 가져야 할까요?
+    - 보통 Util 클래스는 문자열 관련, 랜덤값 구하기, 날짜 및 시간 구하기 등 특정 비지니스 로직이나 독립적인 기능을 한다고 한다.
+    - 책임은 어떤 요청에 대해 그 클래스가 꼭 해야하는 일과 같은 개념인 것 같다. 그렇다면 Util 클래스의 책임은 여러 계층에서 사용될 수도 있고 특별한 의존 관계가 없는 일인 것 같다.
+    - 그 기준을 아직은 정확하게 잡기는 어렵지만 `java.util` , `java.lang` , `java.time` 패키지들을 사용하는 간단한 동작들을 Util 클래스로 만들어 사용할 수 있을 것 같다.
+      예를들어 `java.util.Random`을 사용하여 랜덤값을 만들어내는 경우 `RandomUtil`을 만들어 여러 상황에서의 랜덤값을 얻어올 수 있다. 물론 지금은 `Cars`에서만 자동차 이동을 위해
+      0~9까지의 랜덤값을 받아오는 기능을 사용하고 있어서 Util 클래스를 만들지 않고 `Cars`에 직접 구현할수도 있겠지만, 랜덤 값을 만드는 것이 자동차가 할 일인가? 라고 생각하면 아니라고 생각되어서
+      `RandomUtil`을 만드는 것이 더 좋을 것 같다. 또 `InputView`에서 자동차의 이름을 받아와서 split하는 메서드가 있는데 이런 것도 `StringUtil` 클래스를 만들어 사용해도 될
+      것 같다. 입력을 나누는 일이 `View`의 일인가? 라고 생각하면 아니라고 생각되었기 때문입니다.
+    - `RandomUtil`은 시작 숫자와 끝 숫자만을 받아와 그 사이에 랜덤값을 반환하는 기능을 하는 메서드를 가지고 시작 숫자와 끝 숫자는 Cars가 가지도록 수정.
+    - `MovementUtil`은 파라미터로 받은 값에 따라 이동할지 말지를 확인하고 있는데 이것은 `Car`에서 하도록 수정.
+- [x] 테스트에서만 사용되는 클래스는 test source set에 선언해주면 어떨까요?
 - [x] `Integer.parseInt()`가 중복으로 호출되는곳이 있는데, 어떻게 하면 호출을 줄일 수 있을까요?
-    - Integer.parseInt()는 내부에서 입력이 숫자가 아니면 알아서 NumberFormatException을 던지는 것 같다. 
+    - Integer.parseInt()는 내부에서 입력이 숫자가 아니면 알아서 NumberFormatException을 던지는 것 같다.
     - 그래서 validateNumberFormat 메서드는 필요가 없을 것 같다.
     - 하지만 toInteger에서 NumberFormatException이 발생했을 때 원하는 에러 메시지로 예외를 던지고 싶어서 try catch문을 이용했다.
 - [x] 어떤 예외를 던지는지 다시 확인해보기
@@ -204,10 +213,14 @@
     - CarName에서 validEmpty() 메서드 명도 수정해야한다.
     - 메소드 명, 파라미터 명, 변수 명을 수정
 - [x] toString()를 어떻게 잘 사용해야할지 고민해자!
-    - 수정 전 RoundResult는 내가 정한 규칙으로 반환한 toString()을 이용하여 나만 알수있는 방법으로 데이터를 가져왔던것 같다. 다른 개발자가 이 코드를 본다면 쉽게 이해하기 어려울 수 있을 것 같다. toString()을 활용하는 방법을 잘못이해했기 때문에 발생한 문제인것같다.
-    - "이펙티브 자바의 Item 12"와 여러 블로그에서 toString()에 대해 보면서 toString()은 디버깅을 위해 사용하는 것 외에 사용은 하지않는 것이 좋다고 생각하게 되었다. 그래서 코드 부분과 테스트에서 toString()을 이용하여 테스트하는 모든 부분을 수정하였다.
-    - 처음에 RoundResult에서 toString()으로 데이터를 가져온 이유는 강박적으로 getter를 쓰지 않기위해서 였다. 그런데 조금 더 생각해보니 Car 객체의 이름과 이동 거리를 view에서 사용하게 되었기 때문에 getter를 써도 될것같다고 생각했다.
-    - getter를 쓰지말라는 이유는 getter로 반환된 값을 다른 메서드가 수정하거나 사용하는 로직을 처리하지 말라는 식으로 이해를 했다. 따라서 view에서 사용하기 위해서 getter를 하는것은 허용이 되지않을까? 라고 생각을 했다.
+    - 수정 전 RoundResult는 내가 정한 규칙으로 반환한 toString()을 이용하여 나만 알수있는 방법으로 데이터를 가져왔던것 같다. 다른 개발자가 이 코드를 본다면 쉽게 이해하기 어려울 수 있을 것
+      같다. toString()을 활용하는 방법을 잘못이해했기 때문에 발생한 문제인것같다.
+    - "이펙티브 자바의 Item 12"와 여러 블로그에서 toString()에 대해 보면서 toString()은 디버깅을 위해 사용하는 것 외에 사용은 하지않는 것이 좋다고 생각하게 되었다. 그래서 코드 부분과
+      테스트에서 toString()을 이용하여 테스트하는 모든 부분을 수정하였다.
+    - 처음에 RoundResult에서 toString()으로 데이터를 가져온 이유는 강박적으로 getter를 쓰지 않기위해서 였다. 그런데 조금 더 생각해보니 Car 객체의 이름과 이동 거리를 view에서
+      사용하게 되었기 때문에 getter를 써도 될것같다고 생각했다.
+    - getter를 쓰지말라는 이유는 getter로 반환된 값을 다른 메서드가 수정하거나 사용하는 로직을 처리하지 말라는 식으로 이해를 했다. 따라서 view에서 사용하기 위해서 getter를 하는것은 허용이
+      되지않을까? 라고 생각을 했다.
 - [x] Cars를 생성할 때 차가 0개인 경우 Cars가 생성되지 않도록 수정
     - Cars 생성 시 차가 없으면 `IllegalArgumentException` 발생
     - findWinners()를 실행했을 때 차가 없는 경우 `IllegalStateException` 발생
